src/pyproject.toml
[project]
name = "projektbazydanych"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf2>=2.8.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "psycopg2-binary>=2.9.11",
    "streamlit>=1.52.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
